feat: address issue #68
