+++ packages/cli/src/run-artifacts.ts
@@ -1,0 +1,22 @@
+import { relative } from "node:path";
+
+export function toRepoRelativePath(repoRoot: string, filePath: string): string {
+  return (relative(repoRoot, filePath) || ".").split("\\").join("/");
+}
+
+export function formatRunTimestamp(date = new Date()): string {
+  const pad = (value: number, length = 2): string =>
+    String(value).padStart(length, "0");
+
+  return [
+    `${date.getUTCFullYear()}`,
+    pad(date.getUTCMonth() + 1),
+    pad(date.getUTCDate()),
+    "T",
+    pad(date.getUTCHours()),
+    pad(date.getUTCMinutes()),
+    pad(date.getUTCSeconds()),
+    pad(date.getUTCMilliseconds(), 3),
+    "Z",
+  ].join("");
+}
